daemon: move hubble specific code to a dedicated file
This commit brings no functional changes.
