+++ daemon/cmd/daemon_main.go
@@ -30,7 +30,6 @@ import (
 	"github.com/cilium/cilium/api/v1/server"
 	"github.com/cilium/cilium/api/v1/server/restapi"
 	"github.com/cilium/cilium/common"
-	"github.com/cilium/cilium/pkg/api"
 	"github.com/cilium/cilium/pkg/bpf"
 	"github.com/cilium/cilium/pkg/cgroups"
 	"github.com/cilium/cilium/pkg/cleanup"
@@ -43,9 +42,7 @@ import (
 	"github.com/cilium/cilium/pkg/defaults"
 	"github.com/cilium/cilium/pkg/envoy"
 	"github.com/cilium/cilium/pkg/flowdebug"
-	"github.com/cilium/cilium/pkg/hubble"
 	"github.com/cilium/cilium/pkg/identity"
-	"github.com/cilium/cilium/pkg/ipcache"
 	"github.com/cilium/cilium/pkg/k8s"
 	"github.com/cilium/cilium/pkg/k8s/watchers"
 	"github.com/cilium/cilium/pkg/kvstore"
@@ -64,11 +61,6 @@ import (
 	"github.com/cilium/cilium/pkg/probe"
 	"github.com/cilium/cilium/pkg/sysctl"
 	"github.com/cilium/cilium/pkg/version"
-	hubbleServe "github.com/cilium/hubble/cmd/serve"
-	hubbleMetrics "github.com/cilium/hubble/pkg/metrics"
-	"github.com/cilium/hubble/pkg/parser"
-	hubbleServer "github.com/cilium/hubble/pkg/server"
-	"github.com/cilium/hubble/pkg/server/serveroption"
 	"github.com/go-openapi/loads"
 	gops "github.com/google/gops/agent"
 	"github.com/jessevdk/go-flags"
@@ -1591,55 +1583,4 @@ func checkNodePortAndEphemeralPortRanges() error {
 			ephemeralPortRange[0], ephemeralPortRange[1])
 	}
 	return nil
-}
-
-func (d *Daemon) launchHubble() {
-	logger := logging.DefaultLogger.WithField(logfields.LogSubsys, "hubble")
-	if !option.Config.EnableHubble {
-		logger.Info("Hubble server is disabled")
-		return
-	}
-	addresses := append(option.Config.HubbleListenAddresses, "unix://"+option.Config.HubbleSocketPath)
-	for _, address := range addresses {
-		// TODO: remove warning once mutual TLS has been implemented
-		if !strings.HasPrefix(address, "unix://") {
-			logger.WithField("address", address).Warn("Hubble server will be exposing its API insecurely on this address")
-		}
-	}
-	payloadParser, err := parser.New(d, d, d, ipcache.IPIdentityCache, d)
-	if err != nil {
-		logger.WithError(err).Error("Failed to initialize Hubble")
-		return
-	}
-	s, err := hubbleServer.NewLocalServer(payloadParser, logger,
-		serveroption.WithMaxFlows(option.Config.HubbleFlowBufferSize),
-		serveroption.WithMonitorBuffer(option.Config.HubbleEventQueueSize))
-	if err != nil {
-		logger.WithError(err).Error("Failed to initialize Hubble")
-		return
-	}
-	go s.Start()
-	d.monitorAgent.GetMonitor().RegisterNewListener(d.ctx, hubble.NewHubbleListener(s))
-
-	listeners, err := hubbleServe.SetupListeners(addresses, api.CiliumGroupName)
-	if err != nil {
-		logger.WithError(err).Error("Failed to initialize Hubble")
-		return
-	}
-
-	logger.WithField("addresses", addresses).Info("Starting Hubble server")
-	if err := hubbleServe.Serve(d.ctx, logger, listeners, s); err != nil {
-		logger.WithError(err).Error("Failed to start Hubble server")
-		return
-	}
-	if option.Config.HubbleMetricsServer != "" {
-		logger.WithFields(logrus.Fields{
-			"address": option.Config.HubbleMetricsServer,
-			"metrics": option.Config.HubbleMetrics,
-		}).Info("Starting Hubble Metrics server")
-		if err := hubbleMetrics.EnableMetrics(log, option.Config.HubbleMetricsServer, option.Config.HubbleMetrics); err != nil {
-			logger.WithError(err).Warn("Failed to initialize Hubble metrics server")
-			return
-		}
-	}
 }
